processor/requirements.txt
# NLP Processor dependencies
spacy==3.7.4
psycopg2-binary==2.9.10
python-dotenv==1.1.0
